rut ready - makes debs even better

+++ build.sh
@@ -1,16 +1,17 @@
 #!/bin/bash
 
 ##basic input flags
-if [ $1 = "-h" ]; then
-  echo "builds codebase for this repos .cpp, run as-is or with optional '-a' for clean-All"
-  exit
-elif [ $1 = "-a" ]; then
-  echo "Removing `find ./ -name build-\*| wc -l` build-* folders";
-  rm -rf build-*
-else
-  echo "Starting build with `find ./ -name build-\*| wc -l` build-* folders present"
+if [ -z "$1" ]; then
+    echo "Starting build with `find ./ -name build-\*| wc -l` build-* folders present"
+    else 
+    if [ $1 = "-h" ]; then
+        echo "builds codebase for this repos .cpp, run as-is or with optional '-a' for clean-All"
+        exit
+    elif [ $1 = "-a" ]; then
+        echo "Removing `find ./ -name build-\*| wc -l` build-* folders";
+        rm -rf build-*
+    fi
 fi
-
 
 echo "(Docker entrypoint) running  $0, starting cross-compiled exe and package build(s)."
 echo "Current directory is $PWD"
@@ -77,24 +78,10 @@ ARCH="x86_64"
 export ARCH
 build_for_arch $ARCH "$CROSS_COMPILE_X86_64" "Unix Makefiles"
 
-exit
+#exit
 echo "########################### INFO ############################"
-##debug
-ls -lah ./
-echo "File/folder count for this build dir:" `find ./ | wc -l`
-
-##run test of c++ code in local bash mode
-echo "Running basic test of build outout"
-#####if pwd != /opt/$proj/ then set different thisarch/nonarch
-BLDARCH=build-`arch`
-cd ./$BLDARCH/
-echo "Execute test of compiled c++ code on target host $HOSTNAME running on `arch` $(stat -c "%s%n" -- /.dockerenv 2>/dev/null)"
-echo "`uname -a`"
-
-echo "########################### DEBUG ###########################"
-cd ../
+#cd ../
 NEWESTDEB=`find ./build-aarch64 -type f -name \*.deb | grep CPack | head -n 1`
-#echo `dpkg-deb -c $NEWESTDEB | wc -l` `dpkg-deb -c $NEWESTDEB | grep "goose/goose"`
 echo deb contains `dpkg-deb -c $NEWESTDEB | wc -l` files including `dpkg-deb -c $NEWESTDEB | grep /opt/goose/goose`
 ls -lah $NEWESTDEB
 file ./build-aarch64/$CMPROJ
@@ -107,3 +94,16 @@ file ./build-x86_64/$CMPROJ
 file ./build-x86_64/test/unit_tests
 ./build-x86_64/test/unit_tests
 
+exit
+echo "########################### DEBUG ###########################"
+##debug
+ls -lah ./
+echo "File/folder count for this build dir:" `find ./ | wc -l`
+
+##run test of c++ code in local bash mode
+echo "Running basic test of build outout"
+#####if pwd != /opt/$proj/ then set different thisarch/nonarch
+BLDARCH=build-`arch`
+cd ./$BLDARCH/
+echo "Execute test of compiled c++ code on target host $HOSTNAME running on `arch` $(stat -c "%s%n" -- /.dockerenv 2>/dev/null)"
+echo "`uname -a`"
